Refactor message pane into main:messages.
