+++ process_manager/views/partials.py
@@ -1,14 +1,10 @@
 """View functions for partials."""
 
 import django_tables2
-from django.conf import settings
 from django.contrib.auth.decorators import login_required
 from django.http import HttpRequest, HttpResponse
 from django.shortcuts import render
-from django.utils.timezone import localtime
 from druncschema.process_manager_pb2 import ProcessInstance
-
-from main.models import DruncMessage
 
 from ..process_manager_interface import get_session_info
 from ..tables import ProcessTable
@@ -91,25 +87,4 @@ def process_table(request: HttpRequest) -> HttpResponse:
         request=request,
         context=dict(table=table),
         template_name="process_manager/partials/process_table.html",
-    )
-
-
-@login_required
-def messages(request: HttpRequest) -> HttpResponse:
-    """Search and render Kafka messages from the database."""
-    search = request.GET.get("search", "")
-    records = DruncMessage.objects.filter(
-        topic__regex=settings.KAFKA_TOPIC_REGEX["PROCMAN"], message__icontains=search
-    )
-
-    # Time is stored as UTC. localtime(t) converts this to our configured timezone.
-    messages = [
-        f"{localtime(record.timestamp).strftime('%Y-%m-%d %H:%M:%S')}: {record.message}"
-        for record in records
-    ]
-
-    return render(
-        request=request,
-        context=dict(messages=messages[::-1]),
-        template_name="process_manager/partials/message_items.html",
     )
